warda-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
groq==0.4.0
pandas==2.0.3
numpy==1.24.3
torch==2.0.1
huggingface_hub==0.16.4
sentence-transformers==2.2.2
pymysql==1.1.0
sqlalchemy==2.0.25
pydantic==2.4.2
email-validator==2.0.0.post2